some fixes here and there
